Trace Redaction - Single redact process tree commit
This commit is a merge of multiple smaller commits. The objective
is to remove all process and threads from the process tress that
do not appear in the final package.

Non-redacted processes/threads should appear where they were
originally found, but the synth threads should appear in the
first process tree.

Since a transform primitive cannot change state, a collect and
build primitive need to augment the context so that the transform
primitive knows where to write the synth threads.